+++ transports/changelog.md
@@ -1,5 +1,5 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- fix: handle case when config store is nil in session and plugins handlers
-- chore: adds integration tests for different config combinations
+- feat: Adds option to disable authentication on inference calls
+- chore: Adds dark image for new version infographic